Backend/requirements.txt
Flask
Flask-Cors
numpy
pandas
scikit-learn
pymongo[srv]
python-dotenv
joblib
